e310x: move PLIC implementation to SiFive crate
The PLIC implementation used by the Freedom E310x chips is used by
multiple chips, for instance by the QEMU riscv32 virt-machine type as
well.

As per Alistair Francis' request, this moves the implementation to the
sifive-chip crate accordingly, such that it can be reused by different
chips.

Suggested-by: Alistair Francis <alistair.francis@wdc.com>
